requirements.txt
cookiecutter
GitPython
python-dotenv